created api routes directory and moved user routes into it
