+++ server.js
@@ -24,7 +24,7 @@ app.use(passport.session());
 
 // Requiring our routes
 require('./routes/html-routes.js')(app);
-require('./routes/api-routes.js')(app);
+require('./routes/api/user-routes.js')(app);
 
 // Syncing our database and logging a message to the user upon success
 db.sequelize.sync().then(() => {
